Moved the Structure Map policies to the Infrastructure project. Updated all references and tests. Update the domain project so it is no longer dependent on structure map
