+++ src/SFA.DAS.Commitments.Api.Client/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="FAKEBuildScript" version="1.0.13" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
   <package id="NuGet.CommandLine" version="3.3.0" targetFramework="net45" />
   <package id="NUnit.Console" version="3.4.0" targetFramework="net45" />
   <package id="NUnit.ConsoleRunner" version="3.4.0" targetFramework="net45" />
