+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/packages.config
@@ -6,13 +6,6 @@
   <package id="Microsoft.Data.OData" version="5.7.0" targetFramework="net45" />
   <package id="Microsoft.Data.Services.Client" version="5.7.0" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net45" />
-  <package id="NLog" version="4.3.6" targetFramework="net45" />
-  <package id="SFA.DAS.Configuration" version="1.0.0.6634" targetFramework="net45" />
-  <package id="SFA.DAS.Configuration.AzureTableStorage" version="1.0.0.6634" targetFramework="net45" />
-  <package id="SFA.DAS.Configuration.FileStorage" version="1.0.0.6634" targetFramework="net45" />
-  <package id="SFA.DAS.Messaging" version="2.0.0.6450" targetFramework="net45" />
-  <package id="SFA.DAS.Messaging.AzureServiceBus" version="2.0.0.6450" targetFramework="net45" />
-  <package id="structuremap" version="4.2.0.402" targetFramework="net45" />
   <package id="System.Spatial" version="5.7.0" targetFramework="net45" />
   <package id="WindowsAzure.ServiceBus" version="3.1.7" targetFramework="net45" />
   <package id="WindowsAzure.Storage" version="6.2.0" targetFramework="net45" />
