+++ src/SFA.DAS.EmployerApprenticeshipsService.Infrastructure/DependencyResolution/LoggingPolicy.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Linq;
+using System.Reflection;
 using NLog;
 using StructureMap;
 using StructureMap.Pipeline;
 
-namespace SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution
+namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.DependencyResolution
 {
     public class LoggingPolicy : ConfiguredInstancePolicy
     {
