+++ src/SFA.DAS.EmployerApprenticeshipsService.Infrastructure/DependencyResolution/MessagePolicy.cs
@@ -7,16 +7,16 @@ using SFA.DAS.Configuration;
 using SFA.DAS.Configuration.AzureTableStorage;
 using SFA.DAS.Configuration.FileStorage;
 using SFA.DAS.EmployerApprenticeshipsService.Domain.Attributes;
-using SFA.DAS.EmployerApprenticeshipsService.Domain.Configuration;
+using SFA.DAS.EmployerApprenticeshipsService.Domain.Interfaces;
 using SFA.DAS.Messaging;
 using SFA.DAS.Messaging.AzureServiceBus;
 using SFA.DAS.Messaging.FileSystem;
 using StructureMap;
 using StructureMap.Pipeline;
 
-namespace SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution
+namespace SFA.DAS.EmployerApprenticeshipsService.Infrastructure.DependencyResolution
 {
-    public class MessagePolicy : ConfiguredInstancePolicy
+    public class MessagePolicy<T> : ConfiguredInstancePolicy where T : IConfiguration
     {
         private readonly string _serviceName;
 
@@ -28,7 +28,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution
         protected override void apply(Type pluginType, IConfiguredInstance instance)
         {
             var messagePublisher = instance?.Constructor?
-                    .GetParameters().FirstOrDefault(x => x.ParameterType == typeof(IMessagePublisher) || x.ParameterType == typeof(IPollingMessageReceiver));
+                .GetParameters().FirstOrDefault(x => x.ParameterType == typeof(IMessagePublisher) || x.ParameterType == typeof(IPollingMessageReceiver));
 
             var environment = Environment.GetEnvironmentVariable("DASENV");
             if (string.IsNullOrEmpty(environment))
@@ -38,17 +38,15 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution
 
             if (messagePublisher != null)
             {
-                var queueName = instance
-                                    .SettableProperties()
-                                    .FirstOrDefault(
-                                        c=>c.CustomAttributes
-                                            .FirstOrDefault(x=>x.AttributeType.Name == nameof(QueueNameAttribute)) != null);
+                var queueName = instance.SettableProperties()
+                                .FirstOrDefault(c=>c.CustomAttributes.FirstOrDefault(
+                                                    x=>x.AttributeType.Name == nameof(QueueNameAttribute)) != null);
 
                 if (queueName != null)
                 {
                     var configurationService = new ConfigurationService(GetConfigurationRepository(), new ConfigurationOptions(_serviceName, environment, "1.0"));
 
-                    var config = configurationService.Get<EmployerApprenticeshipsServiceConfiguration>();
+                    var config = configurationService.Get<T>();
                     if (string.IsNullOrEmpty(config.ServiceBusConnectionString))
                     {
                         var queueFolder = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
