+++ src/SFA.DAS.EmployerApprenticeshipsService.Web.UnitTests/Infrastructure/StructureMapRegistrationTests/WhenInitalisingTheContainer.cs
@@ -1,7 +1,7 @@
 ﻿using NUnit.Framework;
 using SFA.DAS.EmployerApprenticeshipsService.Domain.Configuration;
-using SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution;
 using SFA.DAS.EmployerApprenticeshipsService.Domain.Interfaces;
+using SFA.DAS.EmployerApprenticeshipsService.Infrastructure.DependencyResolution;
 using SFA.DAS.EmployerApprenticeshipsService.Web.DependencyResolution;
 using StructureMap;
