+++ src/SFA.DAS.EmployerApprenticeshipsService.Web/DependencyResolution/IoC.cs
@@ -17,7 +17,7 @@
 
 
 using SFA.DAS.EmployerApprenticeshipsService.Domain.Configuration;
-using SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution;
+using SFA.DAS.EmployerApprenticeshipsService.Infrastructure.DependencyResolution;
 
 namespace SFA.DAS.EmployerApprenticeshipsService.Web.DependencyResolution {
     using StructureMap;
@@ -30,7 +30,7 @@ namespace SFA.DAS.EmployerApprenticeshipsService.Web.DependencyResolution {
             {
                 c.Policies.Add(new ConfigurationPolicy<EmployerApprenticeshipsServiceConfiguration>("SFA.DAS.EmployerApprenticeshipsService"));
                 c.Policies.Add<LoggingPolicy>();
-                c.Policies.Add(new MessagePolicy(ServiceName));
+                c.Policies.Add(new MessagePolicy<EmployerApprenticeshipsServiceConfiguration>(ServiceName));
                 c.AddRegistry<DefaultRegistry>();
             });
         }
