+++ src/SFA.DAS.LevyDeclarationProvider.Worker/app.config
@@ -90,7 +90,7 @@
   </appSettings>
   <uri>
     <schemeSettings>
-      <add name="https" genericUriParserOptions="DontUnescapePathDotsAndSlashes"/>
+      <add name="https" genericUriParserOptions="DontUnescapePathDotsAndSlashes" />
     </schemeSettings>
   </uri>
 </configuration>
